+++ src/geodesy/operators/mod.rs
@@ -1,7 +1,7 @@
 use geodesy_rs::authoring::*;
 
 mod unitconvert;
-mod units;
+pub mod units;
 
 #[rustfmt::skip]
 pub const ACCESSORY_OPERATORS: [(&str, OpConstructor); 1] = [
